requirements.txt
pip>=20
django
djangorestframework
uwsgi>=2.0.17.1
uwsgitop